# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Nagios Management Portal API",
  "version": "1.0.8"
 },
 "paths": {
...
  "/api/notification-channels": {
   "get": {
    "tags": [
     "notifications"
    ],
    "summary": "List Channels",
    "operationId": "list_channels_api_notification_channels_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "notifications"
    ],
    "summary": "Create Channel",
    "operationId": "create_channel_api_notification_channels_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChannelIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/notification-channels/{cid}": {
   "put": {
    "tags": [
     "notifications"
    ],
    "summary": "Update Channel",
    "operationId": "update_channel_api_notification_channels__cid__put",
    "parameters": [
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChannelIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "notifications"
    ],
    "summary": "Delete Channel",
    "operationId": "delete_channel_api_notification_channels__cid__delete",
    "parameters": [
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/notification-channels/{cid}/test": {
   "post": {
    "tags": [
     "notifications"
    ],
    "summary": "Test Channel",
    "operationId": "test_channel_api_notification_channels__cid__test_post",
    "parameters": [
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/notification-rules": {
   "get": {
    "tags": [
     "notifications"
    ],
    "summary": "List Rules",
    "operationId": "list_rules_api_notification_rules_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "notifications"
    ],
    "summary": "Create Rule",
    "operationId": "create_rule_api_notification_rules_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/notification-rules/{rid}": {
   "put": {
    "tags": [
     "notifications"
    ],
    "summary": "Update Rule",
    "operationId": "update_rule_api_notification_rules__rid__put",
    "parameters": [
     {
      "name": "rid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Rid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "notifications"
    ],
    "summary": "Delete Rule",
    "operationId": "delete_rule_api_notification_rules__rid__delete",
    "parameters": [
     {
      "name": "rid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Rid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/notifications": {
   "get": {
    "tags": [
     "notifications"
    ],
    "summary": "Feed",
    "operationId": "feed_api_notifications_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "unread_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Unread Only"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ChannelIn": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "provider": {
      "type": "string",
      "title": "Provider"
     },
     "settings": {
      "additionalProperties": true,
      "type": "object",
      "title": "Settings"
     },
     "is_enabled": {
      "type": "boolean",
      "title": "Is Enabled",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "name",
     "provider"
    ],
    "title": "ChannelIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RuleIn": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "event_types": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "minItems": 1,
      "title": "Event Types"
     },
     "channel_id": {
      "type": "integer",
      "title": "Channel Id"
     },
     "throttle_minutes": {
      "type": "integer",
      "maximum": 10080.0,
      "minimum": 0.0,
      "title": "Throttle Minutes",
      "default": 30
     },
     "environments": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Environments"
     },
     "group_ids": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Group Ids"
     },
     "server_ids": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Server Ids"
     },
     "skip_acknowledged": {
      "type": "boolean",
      "title": "Skip Acknowledged",
      "default": true
     },
     "is_enabled": {
      "type": "boolean",
      "title": "Is Enabled",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "name",
     "event_types",
     "channel_id"
    ],
    "title": "RuleIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}